return prerews
